ts: translateError and proxy for unknown errs that have logs
